refactor: reorganize component structure for better maintainability
## Summary
- Reorganized React components into logical directories for better code organization
- Separated page-level components from reusable components
- Established clear naming conventions and file structure patterns

## Changes Made
### New Directory Structure
- src/pages/ - All page-level route components
- src/components/ui/ - Basic UI components (buttons, loading, themes)
- src/components/forms/ - Form-related components
- src/components/layout/ - Layout components (Sidebar)
- src/components/shared/ - Business logic components (MoodCompass, QuestTracker)

### Component Migrations
- Pages (9 components): HomeScreen → HomePage, Chatbot → ChatPage, etc.
- UI Components (8 components): BackButton, Loading, ThemeSelector, etc.
- Forms (1 component): NewPlanForm moved from SchoolSupport/
- Layout (2 components): Sidebar, SidebarItem
- Shared (5 components): MoodCompass, QuestTracker, PanicButton, etc.

### Infrastructure Updates
- Updated App.tsx routing configuration with new component paths
- Fixed all import statements across 50+ files
- Removed duplicate files and empty directories
- Updated lazy-loading imports for better code splitting
- Restored lint-staged configuration in package.json

### Benefits
- Clear separation between pages and reusable components
- Easier component discovery and maintenance
- Consistent naming conventions ([Feature]Page.tsx pattern)
- Better code organization for team development

## Test Plan
- [x] TypeScript compilation passes without errors
- [x] Production build completes successfully
- [x] All import paths resolved correctly
- [x] Component functionality preserved
