+++ src/components/ui/Progress.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 
 type ProgressProps = {
   value: number;
