+++ src/components/ui/SmartLoading.tsx
