--- src/screens/HomeScreen.css
@@ -1,15 +1,0 @@
-.bounce {
-  animation: pop 0.3s ease-out;
-}
-
-@keyframes pop {
-  0% {
-    transform: scale(1);
-  }
-  50% {
-    transform: scale(1.15);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
